+++ src/utils/collection.ts
@@ -1,5 +1,3 @@
-import fs from 'fs'
-import path from 'path'
 import { Collection } from '../types'
 import { BeeArgumentError } from './error'
 import { fileArrayBuffer } from './file'
@@ -17,47 +15,6 @@ export function assertCollection(data: unknown): asserts data is Collection<Uint
   if (!isCollection(data)) {
     throw new BeeArgumentError('invalid collection', data)
   }
-}
-
-/**
- * Creates array in the format of Collection with data loaded from directory on filesystem.
- * The function loads all the data into memory!
- *
- * @param dir path to the directory
- */
-export async function makeCollectionFromFS(dir: string): Promise<Collection<Uint8Array>> {
-  if (typeof dir !== 'string') {
-    throw new TypeError('dir has to be string!')
-  }
-
-  if (dir === '') {
-    throw new TypeError('dir must not be empty string!')
-  }
-
-  return buildCollectionRelative(dir, '')
-}
-
-async function buildCollectionRelative(dir: string, relativePath: string): Promise<Collection<Uint8Array>> {
-  // Handles case when the dir is not existing or it is a file ==> throws an error
-  const dirname = path.join(dir, relativePath)
-  const entries = await fs.promises.opendir(dirname)
-  let collection: Collection<Uint8Array> = []
-
-  for await (const entry of entries) {
-    const fullPath = path.join(dir, relativePath, entry.name)
-    const entryPath = path.join(relativePath, entry.name)
-
-    if (entry.isFile()) {
-      collection.push({
-        path: entryPath,
-        data: new Uint8Array(await fs.promises.readFile(fullPath)),
-      })
-    } else if (entry.isDirectory()) {
-      collection = [...(await buildCollectionRelative(dir, entryPath)), ...collection]
-    }
-  }
-
-  return collection
 }
 
 function makeFilePath(file: File) {
@@ -107,34 +64,4 @@ export function getCollectionSize(fileList: FileList | File[]): number {
   }
 
   return sum
-}
-
-/**
- * Calculate folder size recursively
- *
- * @param dir the path to the folder to check
- * @returns size in bytes
- */
-export async function getFolderSize(dir: string): Promise<number> {
-  if (typeof dir !== 'string') {
-    throw new TypeError('dir has to be string!')
-  }
-
-  if (dir === '') {
-    throw new TypeError('dir must not be empty string!')
-  }
-
-  const entries = await fs.promises.opendir(dir)
-  let size = 0
-
-  for await (const entry of entries) {
-    if (entry.isFile()) {
-      const stats = await fs.promises.stat(path.join(dir, entry.name))
-      size += stats.size
-    } else if (entry.isDirectory()) {
-      size += await getFolderSize(path.join(dir, entry.name))
-    }
-  }
-
-  return size
 }
