+++ test/integration/modules/debug/status.spec.ts
@@ -5,8 +5,9 @@ import {
   SUPPORTED_BEE_VERSION_EXACT,
 } from '../../../../src/modules/debug/status'
 import { beeDebugKy } from '../../../utils'
-import fs from 'fs/promises'
+import fs from 'fs'
 import path from 'path'
+import * as util from 'util'
 
 /**
  * Matches these:
@@ -61,7 +62,7 @@ describe('modules/status', () => {
   })
 
   test('SUPPORTED_BEE_VERSION_EXACT should be same as in package.json', async () => {
-    const file = await fs.readFile(path.join(__dirname, '../../../../package.json'))
+    const file = await util.promisify(fs.readFile)(path.join(__dirname, '../../../../package.json'))
     const packageJson = JSON.parse(file.toString())
 
     expect(SUPPORTED_BEE_VERSION_EXACT).toBe(packageJson.engines.bee)
